datatype

+++ src/pennFAT/FAT.c
@@ -32,7 +32,7 @@ void free_directory_node(dir_node *node) {
     free(node);
 }
 
-FAT* make_fat(char* f_name, uint8_t block_num, uint8_t block_size) {
+FAT* make_fat(char* f_name, uint8_t block_num, uint32_t block_size) {
     if (block_num > 32 || block_num < 1) {
         perror("number of blocks needs to be within 1-32");
         return NULL;
@@ -55,7 +55,7 @@ FAT* make_fat(char* f_name, uint8_t block_num, uint8_t block_size) {
 
     strcpy(res->f_name, f_name);
     res->f_name[len] =  '\0';
-    res->block_num = block_num;
+    res->block_num = (uint8_t)block_num;
 
     if (block_size == 0) {
         res->block_size = 256;
